backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "CodeLingo backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "bcrypt>=4.0,<5.0",
    "python-dotenv>=1.0,<2.0",
    "email-validator>=2.1,<3.0",
    "aiosqlite>=0.20,<1.0",
    "asyncpg>=0.29,<1.0",
    "python-multipart>=0.0.9,<1.0",
    "PyJWT>=2.9,<3.0",
    "httpx>=0.28.1",
    "google-genai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include the backend package from the repo root perspective
packages = ["app"]

[tool.uv]
package = true

